Fixes a race condition when calling stop recording.
Although the encode and stopEncoding function of MediaCodecEncoder are called from the same thread, this does not guarantee that they are executed in the same order. Thus, sometimes processInputBuffer could be called after the MediaCodec.BUFFER_FLAG_END_OF_STREAM flag had been sent to the codec. This caused an error that resulted in mStoppedCompleter never being completed and the call was hung up.